moved tests according to feedback
test_nlevels -> test_integrity.py
unsort tests -> test_sorting.py
duplicates and unique test -> test_unqi_dups.py
